moving from prototype to new component
