+++ homeassistant/components/sensor/geo_rss_incidents.py
@@ -1,8 +1,7 @@
 """
-Generic GeoRSS incident service
-Retrieves current incidents or alerts in GeoRSS format, and shows information
-on incidents filtered by distance to the HA instance's location and grouped
-by category.
+Generic GeoRSS service
+Retrieves current incidents or alerts in GeoRSS format, and shows information on incidents filtered by distance
+to the HA instance's location and grouped by category.
 
 Example configuration:
 
@@ -22,39 +21,29 @@ sensor:
 Sample feeds
 
 Fire
-- NSW Rural Fire Service
-  http://www.rfs.nsw.gov.au/feeds/majorIncidents.xml
-- Qld Rural Fire Service
-  https://www.qfes.qld.gov.au/data/alerts/bushfireAlert.xml
-- ACT Emergency Services Agency
-  http://www.esa.act.gov.au/feeds/currentincidents.xml
-- WA Department of Fire and Emergency Services
-  https://www.emergency.wa.gov.au/data/incident_FCAD.rss
-- Tasmania Fire Service
-  http://www.fire.tas.gov.au/Show?pageId=colBushfireSummariesRss
+- NSW Rural Fire Service: http://www.rfs.nsw.gov.au/feeds/majorIncidents.xml
+- Qld Rural Fire Service: https://www.qfes.qld.gov.au/data/alerts/bushfireAlert.xml
+- ACT Emergency Services Agency: http://www.esa.act.gov.au/feeds/currentincidents.xml
+- WA Department of Fire and Emergency Services: https://www.emergency.wa.gov.au/data/incident_FCAD.rss
+- Tasmania Fire Service: http://www.fire.tas.gov.au/Show?pageId=colBushfireSummariesRss
 
 Earthquake
-- USGS
-  https://earthquake.usgs.gov/earthquakes/feed/v1.0/atom.php
-- BGS
-  http://www.bgs.ac.uk/feeds/worldSeismology.xml
-- Natural Resources Canada
-  http://www.earthquakescanada.nrcan.gc.ca/index-en.php?tpl_region=canada&tpl_output=rss
+- USGS: https://earthquake.usgs.gov/earthquakes/feed/v1.0/atom.php
+- BGS: http://www.bgs.ac.uk/feeds/worldSeismology.xml
+- Natural Resources Canada: http://www.earthquakescanada.nrcan.gc.ca/index-en.php?tpl_region=canada&tpl_output=rss
+
 
 """
-
 import asyncio
-import json
 import logging
+import json
 from datetime import timedelta
 
 import voluptuous as vol
 
-import homeassistant.helpers.config_validation as cv
 from homeassistant.components.sensor import PLATFORM_SCHEMA
-from homeassistant.const import (STATE_UNKNOWN, CONF_SCAN_INTERVAL,
-                                 CONF_UNIT_OF_MEASUREMENT, CONF_NAME,
-                                 CONF_ICON)
+import homeassistant.helpers.config_validation as cv
+from homeassistant.const import (STATE_UNKNOWN, CONF_SCAN_INTERVAL, CONF_UNIT_OF_MEASUREMENT, CONF_NAME, CONF_ICON)
 from homeassistant.helpers.entity import Entity, async_generate_entity_id
 from homeassistant.helpers.event import async_track_time_interval
 
@@ -62,29 +51,28 @@ REQUIREMENTS = ['feedparser==5.2.1', 'haversine==0.4.5']
 
 _LOGGER = logging.getLogger(__name__)
 
-CONF_CATEGORIES = 'categories'
-CONF_RADIUS = 'radius'
-CONF_URL = 'url'
-
-DEFAULT_ICON = 'mdi:alert'
-DEFAULT_NAME = "Incident Information Service"
-DEFAULT_RADIUS_IN_KM = 20.0
-DEFAULT_UNIT_OF_MEASUREMENT = 'Incidents'
-
 DOMAIN = 'geo_rss'
 ENTITY_ID_FORMAT = 'sensor.' + DOMAIN + '_{}'
 INCIDENTS = 'incidents'
+
 MIN_TIME_BETWEEN_UPDATES = timedelta(minutes=5)
+
+CONF_URL = 'url'
+CONF_RADIUS = 'radius'
+CONF_CATEGORIES = 'categories'
+
+DEFAULT_NAME = 'Incident Information Service'
+DEFAULT_RADIUS_IN_KM = 20.0
+DEFAULT_UNIT_OF_MEASUREMEMT = 'Incidents'
+DEFAULT_ICON = 'mdi:alert'
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend({
     vol.Required(CONF_URL): cv.string,
     vol.Optional(CONF_RADIUS, default=DEFAULT_RADIUS_IN_KM): vol.Coerce(float),
     vol.Optional(CONF_NAME, default=None): cv.string,
     vol.Optional(CONF_ICON, default=DEFAULT_ICON): cv.icon,
-    vol.Optional(CONF_CATEGORIES, default=[]): vol.All(cv.ensure_list,
-                                                       [cv.string]),
-    vol.Optional(CONF_UNIT_OF_MEASUREMENT,
-                 default=DEFAULT_UNIT_OF_MEASUREMENT): cv.string,
+    vol.Optional(CONF_CATEGORIES, default=[]): vol.All(cv.ensure_list, [cv.string]),
+    vol.Optional(CONF_UNIT_OF_MEASUREMENT, default=DEFAULT_UNIT_OF_MEASUREMEMT): cv.string,
 })
 
 
@@ -98,37 +86,33 @@ def async_setup_platform(hass, config, async_add_devices, discovery_info=None):
     name = config.get(CONF_NAME)
     icon = config.get(CONF_ICON)
     categories = config.get(CONF_CATEGORIES)
-    interval_in_seconds = config.get(CONF_SCAN_INTERVAL) or timedelta(
-        minutes=5)
+    interval_in_seconds = config.get(CONF_SCAN_INTERVAL) or timedelta(minutes=5)
     unit_of_measurement = config.get(CONF_UNIT_OF_MEASUREMENT)
 
     if None in (home_latitude, home_longitude):
         _LOGGER.error("Latitude or longitude not set in Home Assistant config")
         return False
 
-    _LOGGER.debug("latitude=%s, longitude=%s, url=%s, radius=%s",
-                  home_latitude, home_longitude, url, radius_in_km)
+    _LOGGER.debug("latitude=%s, longitude=%s, url=%s, radius=%s", home_latitude,
+                  home_longitude, url, radius_in_km)
 
-    # Create all sensors based on categories.
+    # create all sensors
     devices = []
     devices_by_category = {}
     if not categories:
-        device = GeoRssServiceSensor(hass, None, [], name, icon,
-                                     unit_of_measurement)
+        device = GeoRssServiceSensor(hass, None, [], name, icon, unit_of_measurement)
         devices.append(device)
         devices_by_category[None] = device
     else:
         for category in categories:
-            device = GeoRssServiceSensor(hass, category, [], name, icon,
-                                         unit_of_measurement)
+            device = GeoRssServiceSensor(hass, category, [], name, icon, unit_of_measurement)
             devices.append(device)
             devices_by_category[category] = device
     async_add_devices(devices)
 
-    # Initialise update service.
-    updater = GeoRssServiceUpdater(hass, home_latitude, home_longitude, url,
-                                   radius_in_km, devices_by_category,
-                                   categories)
+    # initialise access to web resource
+    updater = GeoRssServiceUpdater(hass, home_latitude, home_longitude, url, radius_in_km,
+                                   devices_by_category, categories)
     async_track_time_interval(hass, updater.async_update, interval_in_seconds)
     yield from updater.async_update()
     return True
@@ -137,8 +121,7 @@ def async_setup_platform(hass, config, async_add_devices, discovery_info=None):
 class GeoRssServiceSensor(Entity):
     """Representation of a Sensor."""
 
-    def __init__(self, hass, category, incidents, name, icon,
-                 unit_of_measurement):
+    def __init__(self, hass, category, incidents, name, icon, unit_of_measurement):
         """Initialize the sensor."""
         self.hass = hass
         self._category = category
@@ -195,11 +178,11 @@ class GeoRssServiceSensor(Entity):
 
 
 class GeoRssServiceUpdater:
-    """Provides access to GeoRSS feed and creates and updates UI devices."""
+    """Provides access to GeoRSS and creates and updates UI devices."""
 
     def __init__(self, hass, home_latitude, home_longitude, url,
                  radius_in_km, devices_by_category, filter_by_category):
-        """Initialize the update service."""
+        """Initialize the sensor."""
         self._hass = hass
         self._feed = None
         self._home_latitude = home_latitude
@@ -214,7 +197,7 @@ class GeoRssServiceUpdater:
     @asyncio.coroutine
     def async_update(self, *_):
         import feedparser
-        # Retrieve data from GeoRSS feed.
+        # retrieve data
         self._feed = feedparser.parse(self._url)
         if not self._feed:
             _LOGGER.error("Error fetching feed data from %s", self._url)
@@ -222,53 +205,45 @@ class GeoRssServiceUpdater:
             incidents = []
             _LOGGER.info("%s entri(es) available in feed %s",
                          len(self._feed.entries), self._url)
-            # Filter entries by distance from home coordinates.
+            # filter entries by distance from home
             for entry in self._feed.entries:
-                geometry = None
-                print(entry)
                 if hasattr(entry, 'where'):
-                    geometry = entry.where
+                    distance = self.calculate_distance_to_geometry(entry.where)
                 elif hasattr(entry, 'geo_lat') and hasattr(entry, 'geo_long'):
-                    coordinates = (float(entry.geo_long), float(entry.geo_lat))
-                    geometry = {'type': 'Point', 'coordinates': coordinates}
-                if geometry:
-                    distance = self.calculate_distance_to_geometry(geometry)
+                    coordinates = (float(entry.geo_lat), float(entry.geo_long))
+                    distance = self.calculate_distance_to_coordinates(coordinates)
                 if distance <= self._radius_in_km:
-                    incident = self.create_incident(entry, distance, geometry)
+                    incident = self.create_incident(distance, entry)
                     incidents.append(incident)
             tasks = []
             if self._filter_by_category:
-                # Group incidents by category.
+                # group incidents by category
                 incidents_by_category = {}
                 for incident in incidents:
                     if incident.category in incidents_by_category:
-                        incidents_by_category[incident.category].append(
-                            incident)
+                        incidents_by_category[incident.category].append(incident)
                     else:
                         incidents_by_category[incident.category] = [incident]
-                _LOGGER.debug("Incidents by category: %s",
-                              incidents_by_category)
-                # Set new state (incidents) on devices.
+                _LOGGER.debug("Incidents by category: %s", incidents_by_category)
+                # set new state (incidents) on devices
                 for category in incidents_by_category.keys():
-                    # Update existing device with new list of incidents.
+                    # update existing device with new list of incidents
                     if category in self._devices_by_category:
                         device = self._devices_by_category[category]
                         device.state = incidents_by_category[category]
                         tasks.append(device.async_update_ha_state())
-                _LOGGER.debug("Devices by category: %s",
-                              self._devices_by_category)
+                _LOGGER.debug("Devices by category: %s", self._devices_by_category)
             else:
-                # Add all incidents regardless of category.
+                # add all incidents regardless of category
                 device = self._devices_by_category[None]
                 device.state = incidents
                 tasks.append(device.async_update_ha_state())
-                _LOGGER.debug("Devices by category: %s",
-                              self._devices_by_category)
+                _LOGGER.debug("Devices by category: %s", self._devices_by_category)
             if tasks:
                 yield from asyncio.wait(tasks, loop=self._hass.loop)
 
     @staticmethod
-    def create_incident(feature, distance, geometry):
+    def create_incident(distance, feature):
         category_candidate = None
         if hasattr(feature, 'category'):
             category_candidate = feature.category
@@ -293,7 +268,6 @@ class GeoRssServiceUpdater:
                         id_candidate,
                         pup_date_candidate,
                         summary_candidate,
-                        geometry,
                         distance)
 
     def calculate_distance_to_geometry(self, geometry):
@@ -301,19 +275,19 @@ class GeoRssServiceUpdater:
         if geometry.type == 'Point':
             distance = self.calculate_distance_to_point(geometry)
         elif geometry.type == 'Polygon':
-            distance = self.calculate_distance_to_polygon(
-                geometry.coordinates[0])
+            distance = self.calculate_distance_to_polygon(geometry.coordinates[0])
         else:
             _LOGGER.info("Not yet implemented: %s", geometry.type)
         return distance
 
     def calculate_distance_to_point(self, point):
-        # Swap coordinates to match: (lat, lon).
+        # from shapely.geometry import shape
+        from haversine import haversine
         coordinates = (point.coordinates[1], point.coordinates[0])
         return self.calculate_distance_to_coordinates(coordinates)
 
     def calculate_distance_to_coordinates(self, coordinates):
-        # Expecting coordinates in format: (lat, lon).
+        # from shapely.geometry import shape
         from haversine import haversine
         distance = haversine(coordinates, self._home_coordinates)
         _LOGGER.debug("Distance from %s to %s: %s km", self._home_coordinates,
@@ -321,46 +295,42 @@ class GeoRssServiceUpdater:
         return distance
 
     def calculate_distance_to_polygon(self, polygon):
+        from haversine import haversine
         distance = float("inf")
         # 1. Check if home coordinates are within polygon
         if self.point_in_polygon(self._home_coordinates, polygon):
             distance = 0
         else:
-            # 2. Calculate distance from polygon by calculating the distance
-            #    to each point of the polygon but not to each edge of the
-            #    polygon; should be good enough
+            # 2. Calculate distance from polygon by calculating the distance to each point of the polygon
+            #    but not to each edge of the polygon; should be good enough
             n = len(polygon)
             for i in range(n):
                 polygon_point = polygon[i]
                 coordinates = (polygon_point[1], polygon_point[0])
-                distance = min(distance,
-                               self.calculate_distance_to_coordinates(
-                                   coordinates))
+                distance = min(distance, self.calculate_distance_to_coordinates(coordinates))
         _LOGGER.debug("Distance from %s to %s: %s km", self._home_coordinates,
                       polygon, distance)
         return distance
 
     @staticmethod
     def point_in_polygon(point, polygon):
-        # Source:
-        # http://geospatialpython.com/2011/08/point-in-polygon-2-on-line.html
+        # Source: http://geospatialpython.com/2011/08/point-in-polygon-2-on-line.html
         x = point[0]
         y = point[1]
-        # Check if point is a vertex.
-        if point in polygon:
-            return True
+        # check if point is a vertex
+        if point in polygon: return True
 
-        # Check if point is on a boundary.
+        # check if point is on a boundary
         for i in range(len(polygon)):
+            p1 = None
+            p2 = None
             if i == 0:
                 p1 = polygon[0]
                 p2 = polygon[1]
             else:
                 p1 = polygon[i - 1]
                 p2 = polygon[i]
-            if p1[1] == p2[1] and p1[1] == y and x > min(p1[0],
-                                                         p2[0]) and x < max(
-                p1[0], p2[0]):
+            if p1[1] == p2[1] and p1[1] == y and x > min(p1[0], p2[0]) and x < max(p1[0], p2[0]):
                 return True
 
         n = len(polygon)
@@ -384,15 +354,13 @@ class GeoRssServiceUpdater:
 class Incident(object):
     """Class for storing incidents retrieved."""
 
-    def __init__(self, category, title, guid, pub_date, description, geometry,
-                 distance):
+    def __init__(self, category, title, guid, pub_date, description, distance):
         """Initialize the data object."""
         self._category = category
         self._title = title
         self._guid = guid
         self._pub_date = pub_date
         self._description = description
-        self._geometry = geometry
         self._distance = distance
 
     @property
@@ -410,10 +378,6 @@ class Incident(object):
     @property
     def description(self):
         return self._description
-
-    @property
-    def geometry(self):
-        return self._geometry
 
     @property
     def distance(self):
